Reset expression context just after resetting per tuple context in ExecModifyTable().
Expression context saved in ps_ExprContext for ModifyTable node is
used to process returning clause and on conflict clause. For every
processed tuple it's reset in ExecProcessReturning() and
ExecOnConflictUpdate(). When a query has both RETURNING and ON
CONFLICT clauses, the reset happens twice and the first one of those
might reset memory used by the other. For some reason this doesn't
show up on HEAD, but is apparent when virtual tuple table slots, which
do not copy the datums in its own memory, are used for tuples returned
by RETURNING clause.

This is fix for a query failing in sql/insert_conflict.sql

insert into insertconflicttest as i values (23, 'Avocado') on conflict (key) do update set fruit = excluded.*::text
  returning *;

Ashutosh Bapat, per suggestion by Andres Freund