tts_nvalid in TupleTableSlot is AttrNumber
The type of tts_nvalid is int, which takes at least 4 bytes, whereas
the number of attributes that a tuple can accomodate is restricted by
the maximum value Var::varattno can take. Hence use the same data type
as that of Var::varattno, which is an AttrNumber, for
TupleTableSlot::tts_nvalid.

Ashutosh Bapat, per suggestion by Andres Freund.